Add files via upload
Simple test code.  Tests whether a given point lies within a bounding box.  Returns 4 bits indicating whether the point is (8) away to the left, (4) off to the right, (2) below the bottom or (1) above the top of the box  (2 bits will be set together if the point is diagonally away from one corner of the box, not adjacent to an edge);  if the point lies within the box, the result will be 0.  Works with 16-bit unpacked co-ordinates, which must be in the same system.  Provides an extra entry point which does not zero the result before it begins; thus allowing a second point to be checked, and the combined result used to determine whether a box is entirely within, entirely outside or overlapping another box.